An attempt to fix a crash when BitVector detects an out-of-bounds access, e.g. when parsing a broken message.
This is a problem I long wanted to fix, and yesterday I saw it striking again and decided to bring it up.

The issue is that BitVector implementation uses assert() to check data coming from outside (e.g. during parsing of a message from a mobile). And if data doesn't match expectations - OpenBTS and smqueue just crash.

Below are two function under the question:
https://github.com/fairwaves/openbts-2.8/blob/master/CommonLibs/BitVector.cpp#L71

uint64_t BitVector::peekField(size_t readIndex, unsigned length) const
{
        uint64_t accum = 0;
        char *dp = mStart + readIndex;
        assert(dp+length <= mEnd);
        for (unsigned i=0; i<length; i++) {
                accum = (accum<<1) | ((*dp++) & 0x01);
        }
        return accum;
}

uint64_t BitVector::peekFieldReversed(size_t readIndex, unsigned length) const
{
        uint64_t accum = 0;
        char *dp = mStart + readIndex + length - 1;
        assert(dp<mEnd);
        for (int i=(length-1); i>=0; i--) {
                accum = (accum<<1) | ((*dp--) & 0x01);
        }
        return accum;
}

I see three ways of fixing this, but each has drawbacks or requires huge amount of code refactoring:

 1. Return some sane value (like 0) if the check fails.
 2. Raise exception.
 3. Refactor BitVector to return error as a return value.

IMHO, the 3rd way is the best, but the amount of work is more than I'm willing to contribute.

The 2nd way would follow OpenBTS's style, but catching one more exception here and there would add more mess to the code and it's hard to guarantee that we catch that exception everywhere. Other wise we would crash (again).

To my taste the 1st way is the cleanest in this case. I.e. let the upper layers of message parsing to handle the wrong parsing. They should handle wrong parsing anyway, and if they don't - it's time to fix this.